feat: implement task management module with kanban board, file upload system, and workflow actions

+++ app/(portal)/dashboard/page.tsx
@@ -350,7 +350,7 @@ const STATUS_DOT: Record<string, string> = {
 }
 
 function formatDeadline(deadline: string | null): string {
-  return formatDateDisplay(deadline, { day: "numeric", month: "short" })
+  return formatDateDisplay(deadline, { day: "numeric", month: "long" })
 }
 
 function DashboardContent({
